chore: TypeScriptファイルに変更し、型を定義
